Fix some issues on QeustionPage component

+++ database/migrations/2020_01_02_081902_create_votables_table.php
@@ -17,7 +17,7 @@ class CreateVotablesTable extends Migration
             $table->unsignedBigInteger('user_id');
             $table->unsignedBigInteger('votable_id');
             $table->string('votable_type');
-            $table->tinyInteger('vote')->comment('-1: down vote, 1: up vote');
+            $table->integer('vote')->comment('-1: down vote, 1: up vote');
             $table->timestamps();
             $table->unique(['user_id', 'votable_id', 'votable_type']);
         });
